Flatten graphQL file structure:
