--- Sources/CorePayments/Networking/GraphQL/Common/Models/Extension.swift
@@ -1,4 +1,0 @@
-struct Extension: Codable {
-
-    let correlationId: String
-}
