--- Sources/CorePayments/Networking/GraphQL/FundingEligibility/Models/FundingEligibilityResponse.swift
@@ -1,4 +1,0 @@
-struct FundingEligibilityResponse: Codable {
-
-    let fundingEligibility: FundingEligibility
-}
