--- Sources/CorePayments/Networking/GraphQL/FundingEligibility/Models/SupportedPaymentMethodsTypeEligibility.swift
@@ -1,5 +1,0 @@
-struct SupportedPaymentMethodsTypeEligibility: Codable {
-
-    let eligible: Bool
-    let reasons: [String]?
-}
