+++ Sources/CorePayments/Networking/GraphQL/FundingEligibilityResponse.swift
@@ -1,3 +1,8 @@
+struct FundingEligibilityResponse: Codable {
+
+    let fundingEligibility: FundingEligibility
+}
+
 struct FundingEligibility: Codable {
 
     let venmo: SupportedPaymentMethodsTypeEligibility
@@ -5,4 +10,10 @@ struct FundingEligibility: Codable {
     let paypal: SupportedPaymentMethodsTypeEligibility
     let paylater: SupportedPaymentMethodsTypeEligibility
     let credit: SupportedPaymentMethodsTypeEligibility
+}
+
+struct SupportedPaymentMethodsTypeEligibility: Codable {
+
+    let eligible: Bool
+    let reasons: [String]?
 }
