+++ Sources/CorePayments/Networking/GraphQL/GraphQLError.swift
@@ -3,3 +3,8 @@ public struct GraphQLError: Codable, Error {
     let message: String
     let extensions: [Extension]?
 }
+
+struct Extension: Codable {
+
+    let correlationId: String
+}
