--- Sources/CorePayments/api/models/APIResult.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum APIResult<T> {
-    case success(T)
-    case failure(GraphQLError)
-}
